Use object library target for easy integration
Instead of building either the static or shared library directly, the cmake object library target is used. Either the static or shared library will be built from that which preserves current behavior.
The cmake object library makes it easy to include OpenXLSX as a cmake object library into a parent project which might want to link multiple libraries into either an application or shared library which isn't possible without needing to link the whole archive.