Fix loading of xlsx files without xl/workbook.xml
I have some xlsx files that don't have a reference to `/xl/workbook.xml` in their `[Content_Types].xml`. OpenXLSX fails to open those, throwing a "path does not exist in zip archive". This fix checks if the `/xl/workbook.xml` reference is in `[Content_Types].xml` and adds it if it isn't.